+++ server/app/src/main/java/eu/europa/ec/eudi/signer/rssp/common/config/OID4VPConfig.java
@@ -1,11 +1,15 @@
 package eu.europa.ec.eudi.signer.rssp.common.config;
 
+import jakarta.validation.Valid;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.PropertySource;
+import org.springframework.validation.annotation.Validated;
 
 @PropertySource("file:application.yml")
 @ConfigurationProperties(prefix = "oid4vp")
+@Validated
 public class OID4VPConfig {
+	@Valid
 	private VerifierProperties verifier;
 	private WalletConfig wallet;